move apply changes to separate package, and rename baseStruct pre-update, remove optional apply changes
