+++ pkg/applychanges/apply_changes.go
@@ -1,4 +1,4 @@
-package resolvers
+package applychanges
 
 import (
 	"reflect"
