+++ pkg/applychanges/apply_changes_test.go
@@ -1,4 +1,4 @@
-package resolvers
+package applychanges
 
 import (
 	"testing"
